backend/requirements.txt
asgiref==3.7.2
django-cors-headers==4.3.1
drf_yasg==1.21.4
drf_extra_fields==3.7.0
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
webcolors==1.11.1
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
python-dotenv==1.0.0
pytz==2022.06
psycopg2-binary==2.9.6
requests==2.26.0
flake8
django-filter==23.1
PyJWT==2.1.0
reportlab
isort==5.12.0
nano==0.10.0
